Added Table setup and Sanitized Input
Added 'create table' feature, to easily add the new table.
Implemented sanitized input for adding a user to the table. This will
also check if the input lengths is within specification. Before the user
is added to the table, MySQLi escapes the characters which helps to
protect against SQL Injection.